refactor auth router for login/register, converted checkJWT to middleware
