+++ api/server.js
@@ -8,6 +8,7 @@ const server = express();
 
 // Library Middleware
 server.use(cors(), helmet(), express.json());
+const authenticate = require("../middleware/auth-middleware");
 
 // Routers
 const usersRouter = require("../controllers/user");
@@ -20,7 +21,7 @@ const errorHandler = require("../middleware/errorHandling");
 // API endpoints
 server.use("/users", usersRouter);
 server.use("/auth", authRouter);
-server.use('/allegiances', allegiancesRouter);
+server.use("/allegiances", allegiancesRouter);
 
 // sanity check
 server.get("/", (req, res) => {
